refactor: service에서 사용하는 repository를 인터페이스로 변경
